brokers_tests/requirements.txt
# Python dependencies for PGMQ Relay integration tests

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Database
psycopg2-binary>=2.9.9

# Kafka client
kafka-python>=2.0.2

# RabbitMQ client
pika>=1.3.2

# NATS client
nats-py>=2.7.0

# Utilities
asyncio>=3.4.3
